feat!: remove HTEnergyBattery and HTEnergyHandler then replace with HTEnergyStorage

+++ src/main/kotlin/hiiragi283/ragium/common/variant/HTMachineVariant.kt
@@ -35,7 +35,7 @@ enum class HTMachineVariant(val tier: HTMachineTier, private val enPattern: Stri
     SIMULATOR(HTMachineTier.ELITE, "Simulation Chamber", "シミュレーション室"),
     ;
 
-    val energyCapacity: Long get() = RagiumConfig.COMMON.machineEnergyCapacity[this]!!.asLong
+    val energyCapacity: Int get() = RagiumConfig.COMMON.machineEnergyCapacity[this]!!.asInt
     val energyUsage: Int get() = RagiumConfig.COMMON.machineEnergyUsage[this]!!.asInt
 
     override val blockHolder: HTDeferredBlock<HTHorizontalEntityBlock, HTMachineBlockItem> by lazy { RagiumBlocks.MACHINES[this]!! }
